refactor: rewrite state computation and update grammar
The grammar has now changed in that replacements and deletions can be
relative to "loci", and that mutations (renamed to "changes") and
plasmids can no longer have "markers" assigned to them.

The optional locus is a new way to resolve ambiguity in changes.

All code is also annotated using type hints located in *.pyi files